project-slm/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
ollama>=0.4.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
duckduckgo-search>=4.0
pyyaml>=6.0
python-frontmatter>=1.0
pydantic>=2.0
httpx>=0.25.0
